Change implementation of model
I removed a Manager class in favor of specific class for model,
in this case is QAbstracItemModel and every time a lang is added
refresh it internal data. I used this instead of QTableModel because
the data are the same after file is imported. After this I will restore
some functions over the model like add/remove lang. Maybe in the future
should be bettere change the structure of data Language-Key.

Signed-off-by: Federico Guerinoni <guerinoni.federico@gmail.com>